Create Lucky Palin (LUCKPAL)
simple logic we have to change the half of the string and make it equal to other half. we cant have lucky written palindrome if size is less then 9.
take every index from first half and try to make it lucky from this index and calculate the total number of operation will required to do this and again try with "ykcul" too. 
and update the answer if number of operation is minimum.