Create My Fair Coin(CSUMD)
for this code to understand you need to first understand two things first is matrix exponent and other is binaryPower. 
if you know these two things it is simple total number of ways of any value of n can be calculated by this logic. if 2 cent place at first position then number of coins are dp[n-2] and if 1 cent is place at 1st position number of coins will be dp[n-1] but you have to multiply it with two(2) as there first position number is only calculated once(only head).