Create One Dimensional Kingdoms
1. observation
f you read the question carefully you'll get that we just have to find the common points between range and if there then number of bomb will decrease.
2. approach and algo.
the approach is basic just sort the array according to increasing order of start point. now if there is no common point the total number of bomb will be 'n' so initializing the size be 'n'. and check every consecutive val if the start of the next value is less than or equal to our end value just decrease the bomb size and update the value of end.