Create Meet In The Median
for the greatest median just sort the array in non increasing order and take first 'k' elements its middle element will be the greatest. now but we have to also write it according to index just take other array of pair for index and value and sort according to index and then print it.