Create Timme measure
after observing the question you could understand you have to find every angle and then check it fits or not if is fit print it.
running two for loop for hour and minute hand. taking angle for both hand and finding the difference if the angle difference is greater than 180 subtracting with 360. now checking the difference between calculate and given value is less thwn1/120 or not if its then simply print it.