Create Too many candies
we just have to throw(subtract) value until n>m and check if we got the n%m==0 then print ans or else '-1'; if observe we just have to check after dividing n from n the remainder should be divisible by k. so we are checking that and if remainder is not divisible we are decrementing number of times in which n was divisible. like for n=26 m=5 k=2; times=(26/5)=5 rem=26-(55)=1 if(1%k==0) that means we can distribute candies equally. or decrease times--(5-1=4) rem=26-(54)=6 now 6 is divisible by k(=2) ans will be rem/k(=3)