Create K-Flip
if you observe the question carefully you'll realize we have to change all 1's that are coming before n-k+1 index, to '0' and the last numbers they are going to be '0' or '1' according to their previous numbers flips.

so what i did i created an boolean flag that will tell we have to flip the value or not. but there is a catch when we use flag as we don't know up to what index i have to flip. so i have created an array of boolean, an whenever the value is 1,I am flipping the sub string and storing up to what index i have to flip. and the second part is going to be according to that array.