Create Beautiful Array
The only thing that matters is the numbers mod 4. So let's just store the numbers mod 4 and the count of each residue.

c[i] = count of numbers in the list that are i (mod 4).
First deal with the numbers that are 1 mod 4 and 3 mod 4. We can repeatedly pair them up to get a sum that's a multiple of 4. This takes min(c[1], c[3]) steps.
After pairing them up, if c[1] or c[3] is odd, then the sum of all the numbers is odd so we can never make all the numbers divisible by 4.

(The other one will be 0).